chore: Readability and comprehensibility of code

+++ scripts/lightweight_curator.py
@@ -12,6 +12,7 @@ import sys
 elasticsearch_host = os.getenv("ELASTICSEARCH_HOST", "elasticsearch:9200")
 percentage_threshold = int(os.getenv("PERCENTAGE_THRESHOLD", "80"))
 index_name_prefixes = os.getenv("INDEX_NAME_PREFIXES", "infra-,app-,audit-")
+dry_run = bool(os.getenv("DRY_RUN", "false"))
 
 # types of logs
 log_info = "info"
@@ -84,7 +85,7 @@ def get_max_allowed_size(es, percentage_threshold):
     for node in data.splitlines():
         i = i + int(node)
 
-    max_allowed_size = (percentage_threshold * i) / 100.0
+    max_allowed_size = int( (percentage_threshold * i) / 100.0 )
 
     return max_allowed_size
 
@@ -94,37 +95,49 @@ def get_first_item(a_dict={}):
     first_value = next(value_iterator)
     return first_value
 
-def get_actionable_indices(es, max_allowed_size, index_name_prefixes_list):
+def indices_smaller_then_max_allowed_size(index, limit, indices_size_counter, indices_to_delete):
     """
-    Returns a list of actionable indices based on percentage size and age.
+    Returns list of indices which are above threshold limit.
     """
-    class data_structure:
+    expected_size = index.size + indices_size_counter
+
+    if indices_size_counter < limit and expected_size < limit:
+        indices_size_counter += index.size
+        log(log_info, f"Do not add into actionable list: {index.name}, summed disk usage is {indices_size_counter} B and disk limit is {limit} B")
+    else:
+        log(log_info, f"Add into actionable list: {index.name}, summed disk usage is {indices_size_counter} B and disk limit is {limit} B")
+        indices_to_delete.append(index.name)
+
+    return indices_to_delete
+
+def get_actionable_indices(es, max_allowed_size, index_name_prefixes):
+    """
+    This function returns a list of indices which will be used in deletion process.
+    """
+    class index_struct:
       def __init__(self, name, size, creation_date):
         self.name = name
         self.size = size
         self.creation_date = creation_date
 
-    # Prepare dictionary of indices with their size and creation_date values.
-    size_counter = 0
-    data_array = []
-
-    for index_name_prefixes in index_name_prefixes_list:
-        for name in es.indices.get_alias(index=index_name_prefixes + "*").keys():
+    """
+    Appends index into the list of indices with their name, size and creation_date.
+    """
+    indices = []
+    for index_name_prefix in index_name_prefixes:
+        for name in es.indices.get_alias(index=index_name_prefix + "*").keys():
             size = int(get_first_item(es.indices.stats(index=name)['indices'][name]['total']['store']))
             creation_date = int(es.indices.get(index=name)[name]['settings']['index']['creation_date'])
-            data_array.append(data_structure(name, size, creation_date))
+            indices.append(index_struct(name, size, creation_date))
 
-    # Output are one or more indices which are above the 80% threshold and are supposed to be deleted.
+    """
+    Iterates through sorted indices (using reverse=True meaning oldest indices will be deleted first)
+    and calculates if index is smaller or bigger then max allowed size.
+    """
     indices_to_delete = []
-    for data_object in sorted(data_array, key=lambda x: x.creation_date, reverse=True):
-        if size_counter < max_allowed_size and data_object.size + size_counter < max_allowed_size:
-            log(log_info, "Do not add into actionable list: '{indice}', summed disk usage is {usage} B and disk limit is {limit} B".format(
-                indice=data_object.name, usage=size_counter, limit=int(max_allowed_size)))
-            size_counter += data_object.size
-        else:
-            log(log_info, "Add into actionable list: '{indice}', summed disk usage is {usage} B and disk limit is {limit} B".format(
-                indice=data_object.name, usage=size_counter, limit=int(max_allowed_size)))
-            indices_to_delete.append(data_object.name)
+    indices_size_counter = 0
+    for index in sorted(indices, key=lambda x: x.creation_date, reverse=True):
+        indices_smaller_then_max_allowed_size(index, max_allowed_size, indices_size_counter, indices_to_delete)
 
     return indices_to_delete
 
@@ -133,12 +146,12 @@ def delete_indices(es, indices_to_delete):
     Delete actionable indices pasted from get_actionable_indices() function.
     """
     # Delete indices.
-    for indice in indices_to_delete:
+    for index in indices_to_delete:
         try:
-            es.indices.delete(index=indice)
-            log(log_info, "Deleted indice '{s}'".format(s=indice))
+            es.indices.delete(index=index)
+            log(log_info, f"Deleted index '{index}'")
         except Exception as e:
-            log(log_err, "Error deleting indice '{s}'".format(s=indice), extra={
+            log(log_err, f"Error deleting index '{index}'", extra={
                 "exception": e
             })
 
@@ -148,27 +161,28 @@ def main():
     global index_name_prefixes
     global elasticsearch_host
     global percentage_threshold
+    global dry_run
 
     # Initial validation of environment variables.
     env_validation(index_name_prefixes, elasticsearch_host)
 
     # Index name prefixes from comma-separated string.
-    index_name_prefixes_list = index_name_prefixes.split(',')
+    index_name_prefixes = index_name_prefixes.split(',')
 
     # Pass elasticsearch connect arguments.
     es = es_connect_args(elasticsearch_host)
 
-    log(log_info, "Removing indices which are above {percentage} threshold from host '{host}'".format(
-        host=elasticsearch_host,
-        percentage=percentage_threshold
-    ))
+    log(log_info, f"""Searching through indices sorted by the age to find and remove first oldest index which exceeds total storage threshold,
+    Value for total storage threshold is set to {percentage_threshold}%,
+    Host name is {elasticsearch_host}""")
 
     # Get list of actionable indices.
-    indices_to_delete = get_actionable_indices(es, get_max_allowed_size(es, percentage_threshold), index_name_prefixes_list)
+    indices_to_delete = get_actionable_indices(es, get_max_allowed_size(es, percentage_threshold), index_name_prefixes)
 
     # For development purpose.
-    print(indices_to_delete)
-    sys.exit(1)
+    if dry_run:
+        print(indices_to_delete)
+        sys.exit(1)
 
     # Delete actionable indices.
     delete_indices(es, indices_to_delete)
